+++ hw-1/src/main/java/ru/dankoy/core/domain/AnswerImpl.java
@@ -1,4 +1,4 @@
-package ru.dankoy.otus.core.domain;
+package ru.dankoy.core.domain;
 
 
 public class AnswerImpl implements Answer {
