+++ hw-1/src/main/java/ru/dankoy/core/service/QuestionsService.java
@@ -1,7 +1,7 @@
-package ru.dankoy.otus.core.service;
+package ru.dankoy.core.service;
 
 import java.util.List;
-import ru.dankoy.otus.core.domain.Question;
+import ru.dankoy.core.domain.Question;
 
 /**
  * @author turtality
